Add sessions store support.
Right now the sessions will be stored on the database and automatically
expire whenever a set amount of time has passed. This is better than just
relying on cookies for session storage since now if the server restarts
the sessions won't all expire.

Still need to make the authentication work across all routes. Currently
just putting the login in app.js for testing.