Add ability to sync/download zip files.
The way it works is through a handshake. There are certain status codes
the server returns depending on how the request to sync has ended.

I don't like the way the file system management is happening for audio,
images and zip files. Need to think of a better more logical way to organize
them.